seperating files. the app.py includes the basic structure while the choices and the choice menu was put into the choice_menu.py
